Best opti ever

+++ includes/errors.h
@@ -6,7 +6,7 @@
 /*   By: mfeldman <mfeldman@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/24 15:17:50 by fllanet           #+#    #+#             */
-/*   Updated: 2024/02/29 08:10:41 by mfeldman         ###   ########.fr       */
+/*   Updated: 2024/03/07 00:46:20 by mfeldman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,22 +25,22 @@ typedef struct s_error
 # define COLOR_RESET   "\x1b[0m"
 # define COLOR_RED     "\x1b[31m"
 
-# define ERROR_ARG				0x1
-# define ERROR_FILE				0x2
-# define ERROR_CUB				0x4
-# define ERROR_EMPTY_S			0x8
-# define ERROR_EMPTY_LINE		0x10
-# define ERROR_SCENE			0X20
-# define ERROR_MAP				0X40
-# define ERROR_RM_MAP			0x80
-# define ERROR_SORT_S			0x100
-# define ERROR_CHAR				0x200
-# define ERROR_START			0x400
-# define ERROR_WALL				0x800
-# define ERROR_XPM				0x1000
-# define ERROR_RGB				0x2000
-# define ERROR_INIT_MLX			0x4000
-# define ERROR_INIT_IMG			0x8000
+# define ERROR_ARG				(uint8_t)0x1
+# define ERROR_FILE				(uint8_t)0x2  
+# define ERROR_CUB				(uint8_t)0x4
+# define ERROR_EMPTY_S			(uint8_t)0x8
+# define ERROR_EMPTY_LINE		(uint8_t)0x10
+# define ERROR_SCENE			(uint8_t)0X20
+# define ERROR_MAP				(uint8_t)0X40
+# define ERROR_RM_MAP			(uint8_t)0x80
+# define ERROR_SORT_S			(uint16_t)0x100
+# define ERROR_CHAR				(uint16_t)0x200
+# define ERROR_START			(uint16_t)0x400
+# define ERROR_WALL				(uint16_t)0x800
+# define ERROR_XPM				(uint16_t)0x1000
+# define ERROR_RGB				(uint16_t)0x2000
+# define ERROR_INIT_MLX			(uint16_t)0x4000
+# define ERROR_INIT_IMG			(uint16_t)0x8000
 # define ERROR_MLX_WIN			0x10000
 # define ERROR_POS				0x20000
 # define ERROR_MALLOC			0x40000
